Resolve issues from fixing resilio sync
