+++ pydiffexp/__init__.py
@@ -1,1 +1,4 @@
-
+from .diffexp import *
+from .plot import *
+from .analyze import *
+from .pipeline import *
